chore: fixed affiliate ib page
