+++ crowdin/messages.json
@@ -3952,6 +3952,7 @@
   "-1943991243": "Partnership with a trusted pioneer",
   "-1781913728": "Benefit from our extensive experience of more than 20 years and our internationally acclaimed reputation.",
   "-1994716313": "No charges or hidden fees",
+  "-1669980379": "Sign up as our affiliate",
   "-13699993": "Deriv Affiliate Programme",
   "-806909894": "Partner with us as an affiliate. Earn commission from your referred client’s deposits.",
   "-1067164259": "Partner with us as an affiliate. Earn commission from the total net revenue of your referred clients’ trades on Deriv Trader, Deriv Bot, and SmartTrader.",
@@ -4071,6 +4072,11 @@
   "-1664474688": "Fill out the online application form. We’ll review your application and get in touch once it’s approved.",
   "-1253892742": "Advertise",
   "-2139211466": "Use your unique affiliate link and our tried-and-tested referral tools to bring new clients to Deriv.",
+  "-871158567": "Deriv Affiliate Programme is free. Enjoy reliable payouts by getting your affiliate commissions paid to your Deriv account.",
+  "-903553712": "Start off as an affiliate and get ensuing access to our IB programme. Earn commission for as long as your referred clients keep trading.",
+  "-2090447319": "DMT5 and Dtrader trading platform at Deriv",
+  "-1299009682": "Optimise conversions with great user experience and creative support",
+  "-1657692765": "Marketing creative material",
   "-2096939984": "Why partner with us",
   "-1379332802": "Earn based on each successful referral.",
   "-1706697508": "This plan is available exclusively for EU-based clients. Please note that according to regulations, you cannot have clients who reside in Portugal or Spain.",
@@ -4082,12 +4088,6 @@
   "-53651607": "Disclaimer: Affiliates not residing in the EU may sign up for the CPA plan.",
   "-704590302": "You earn <0>USD 100</0> when your new referred client deposits and trades a total amount of <0>USD 100</0> or its equivalent into their Deriv account, either in one deposit or cumulatively.",
   "-1794572884": "Affiliates not residing in the EU may sign up for the CPA plan.",
-  "-1669980379": "Sign up as our affiliate",
-  "-871158567": "Deriv Affiliate Programme is free. Enjoy reliable payouts by getting your affiliate commissions paid to your Deriv account.",
-  "-903553712": "Start off as an affiliate and get ensuing access to our IB programme. Earn commission for as long as your referred clients keep trading.",
-  "-2090447319": "DMT5 and Dtrader trading platform at Deriv",
-  "-1299009682": "Optimise conversions with great user experience and creative support",
-  "-1657692765": "Marketing creative material",
   "-1201069751": "Master Affiliate",
   "-1625742908": "This plan is not available for affiliates who promote to clients residing in the EU.",
   "-82217352": "Sub affiliate’s total commission: USD 300",
