+++ src/pages/partners/affiliate-ib/_why-us.tsx
@@ -6,6 +6,7 @@ import { Header, Text, QueryImage } from 'components/elements'
 import { Localize, localize } from 'components/localization'
 import device from 'themes/device'
 import { TString } from 'types/generics'
+import useRegion from 'components/hooks/use-region'
 
 export type WhyUsType = {
     title: TString
@@ -107,8 +108,53 @@ const Content = styled.div`
     }
 `
 
-const WhyUs = ({ items }: WhyUsProps) => {
+type contentType = {
+    [T: string]: TString
+}
+
+const WhyUs = () => {
     const data = useStaticQuery(query)
+    const { is_eu } = useRegion()
+
+    const content_data: contentType = is_eu
+        ? {
+              payout_title: '_t_Zero charges with prompt monthly payouts_t_',
+              payout_text:
+                  '_t_Deriv Affiliate Programme is free. Enjoy reliable payouts by getting your affiliate commissions paid to your Deriv account._t_',
+          }
+        : {
+              payout_title: '_t_Zero charges with prompt monthly and daily payouts_t_',
+              payout_text:
+                  '_t_All Deriv partnership programmes are free. Get your affiliate commissions paid to your method of choice every month and IB commissions paid to your Deriv MT5 account daily._t_',
+          }
+
+    const why_partner_with_us_items: WhyUsType = [
+        {
+            title: '_t_Multiple income opportunities and generous commissions_t_',
+            subtitle:
+                '_t_Start off as an affiliate and get ensuing access to our IB programme. Earn commission for as long as your referred clients keep trading._t_',
+            image_name: 'multiple_income_opportunities',
+            image_alt: '_t_DMT5 and Dtrader trading platform at Deriv_t_',
+            only_row: true,
+        },
+        {
+            title: content_data.payout_title,
+            subtitle: content_data.payout_text,
+            image_name: is_eu ? 'daily_ib_commission_eu' : 'daily_ib_commission',
+            image_alt: '_t_Check your daily IB commission_t_',
+        },
+        {
+            title: '_t_Optimise conversions with great user experience and creative support_t_',
+            subtitle:
+                '_t_We’ve designed a customer-centric and intuitive trading experience for Deriv that’s optimised to convert visitors into clients. We’ll also provide you with the tools and creative materials you need to drive traffic to Deriv._t_',
+            image_name: 'marketing_material',
+            image_alt: '_t_Marketing creative material_t_',
+        },
+    ]
+
+    const why_partner_data = is_eu
+        ? why_partner_with_us_items.filter((item) => !item.only_row)
+        : why_partner_with_us_items
 
     return (
         <StyledContainer>
@@ -116,7 +162,7 @@ const WhyUs = ({ items }: WhyUsProps) => {
                 <WhyUsHeader as="h2" size="4.8rem" align="center">
                     <Localize translate_text="_t_Why partner with us_t_" />
                 </WhyUsHeader>
-                {items.map((item, index) => {
+                {why_partner_data.map((item, index) => {
                     const is_even = index % 2
                     return (
                         <Row flex_direction={is_even ? 'row-reverse' : 'row'} key={index}>
